new game mode -> Done
todo:
3 routes for NIGHT

+++ src/components/MemberWithRole/MemberWithRole.js
@@ -1,46 +1,60 @@
 import React from 'react';
 import css from "../../pages/RolePage/RolePage.module.css";
 
+
 const MemberWithRole = ({member}) => {
 
     const choosenCitizen = (e) => {
         e.preventDefault()
-        const membersCheck = JSON.parse(localStorage.getItem('members')) || []
-        for (const memberCheck of membersCheck) {
-            if (memberCheck.name === member.name) {
-                memberCheck.role = 'citizen'
-                localStorage.setItem('members', JSON.stringify(membersCheck))
+        const resetConfirm = window.confirm("are you sure ?")
+        if (resetConfirm) {
+            const membersCheck = JSON.parse(localStorage.getItem('members')) || []
+            for (const memberCheck of membersCheck) {
+                if (memberCheck.name === member.name) {
+                    memberCheck.role = 'citizen'
+                    localStorage.setItem('members', JSON.stringify(membersCheck))
+                }
             }
-        }
 
+        }
     }
     const choosenMafia = (e) => {
         e.preventDefault()
-        const membersCheck = JSON.parse(localStorage.getItem('members')) || []
-        for (const memberCheck of membersCheck) {
-            if (memberCheck.name === member.name) {
-                memberCheck.role = 'mafia'
-                localStorage.setItem('members', JSON.stringify(membersCheck))
+        const resetConfirm = window.confirm("are you sure ?")
+        if (resetConfirm) {
+            const membersCheck = JSON.parse(localStorage.getItem('members')) || []
+            for (const memberCheck of membersCheck) {
+                if (memberCheck.name === member.name) {
+                    memberCheck.role = 'mafia'
+                    localStorage.setItem('members', JSON.stringify(membersCheck))
+                }
             }
         }
     }
     const choosenDoctor = (e) => {
         e.preventDefault()
-        const membersCheck = JSON.parse(localStorage.getItem('members')) || []
-        for (const memberCheck of membersCheck) {
-            if (memberCheck.name === member.name) {
-                memberCheck.role = 'doctor'
-                localStorage.setItem('members', JSON.stringify(membersCheck))
+        const resetConfirm = window.confirm("are you sure ?")
+        if (resetConfirm) {
+            const membersCheck = JSON.parse(localStorage.getItem('members')) || []
+            for (const memberCheck of membersCheck) {
+                if (memberCheck.name === member.name) {
+                    memberCheck.role = 'doctor'
+                    localStorage.setItem('members', JSON.stringify(membersCheck))
+                }
             }
         }
     }
     const choosenPolice = (e) => {
         e.preventDefault()
-        const membersCheck = JSON.parse(localStorage.getItem('members')) || []
-        for (const memberCheck of membersCheck) {
-            if (memberCheck.name === member.name) {
-                memberCheck.role = 'police'
-                localStorage.setItem('members', JSON.stringify(membersCheck))
+        const resetConfirm = window.confirm("are you sure ?")
+        if (resetConfirm) {
+
+            const membersCheck = JSON.parse(localStorage.getItem('members')) || []
+            for (const memberCheck of membersCheck) {
+                if (memberCheck.name === member.name) {
+                    memberCheck.role = 'police'
+                    localStorage.setItem('members', JSON.stringify(membersCheck))
+                }
             }
         }
     }
